requirements.txt
requests>=2.0.0
attrs>=15.0
